+++ docs/changelog.md
@@ -6,7 +6,7 @@
 
 - Various bug fixes.
 - Improved `repr(layer)`.
-- Added `layer.delete` and `layer.copyto(parent)`.
+- Added `layer.delete` and `layer.copyto(parent)`, and thereby added the ability to set `layer.parent` to be in a different `PXDFile`.
 - Tightened up `layer.mask` behaviour.
 
 ### 0.0.3